Rename domain participant impl file
