+++ dds/src/implementation/dds_impl/mod.rs
@@ -4,9 +4,10 @@ pub mod any_topic_listener;
 pub mod dcps_service;
 pub mod dds_data_reader;
 pub mod dds_data_writer;
+pub mod dds_domain_participant;
 pub mod dds_publisher;
 pub mod dds_subscriber;
-pub mod domain_participant_impl;
+pub mod dds_topic;
 pub mod iterators;
 pub mod message_receiver;
 pub mod node_builtin_data_reader_stateful;
@@ -26,4 +27,3 @@ pub mod node_user_defined_topic;
 pub mod participant_discovery;
 pub mod status_condition_impl;
 pub mod status_listener;
-pub mod dds_topic;
