+++ dds/src/implementation/dds_impl/node_user_defined_data_reader.rs
@@ -28,7 +28,7 @@ use super::{
     any_data_reader_listener::AnyDataReaderListener,
     dcps_service::DcpsService,
     dds_data_reader::DdsDataReader,
-    domain_participant_impl::{AnnounceKind, DdsDomainParticipant},
+    dds_domain_participant::{AnnounceKind, DdsDomainParticipant},
     node_user_defined_subscriber::UserDefinedSubscriberNode,
     node_user_defined_topic::UserDefinedTopicNode,
     status_condition_impl::StatusConditionImpl,
