+++ dds/src/implementation/dds_impl/node_user_defined_subscriber.rs
@@ -28,7 +28,7 @@ use super::{
     dcps_service::DcpsService,
     dds_data_reader::DdsDataReader,
     dds_subscriber::DdsSubscriber,
-    domain_participant_impl::{AnnounceKind, DdsDomainParticipant},
+    dds_domain_participant::{AnnounceKind, DdsDomainParticipant},
     node_domain_participant::DomainParticipantNode,
     node_user_defined_data_reader::UserDefinedDataReaderNode,
     status_listener::StatusListener,
